+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Added
 -----
 * In learn.py, DFData does train/test split.
 * NN Trainer does validation on holdout test data set.
+* Can now transfer Study data to an archive dict.
 
 Fixed
 -----